feat: post-release script changes
BREAKING CHANGE: after a release, master branch will be updated with the
whole release branch, not just the last commit.

If the alpha branch merged to release branch, the changes will already be
present on master branch. If a hotfix branch is directly merged into
release branch, though, master branch will be updated with the hotfix changes.